requirements.txt
requests==2.31.0
python-dotenv==0.9.1
python-telegram-bot==13.2
google_cloud_dialogflow==2.28.2